\name{results}
\alias{results}
\docType{data}
\title{
results of univariate Cox proportional hazard analysis of patients with ADC in
three cohorts.
}
\description{
The "results" is a list consisting of three results generated by performing
univariate Cox proportional hazard survival analysis in three different cohorts.
Each result has four columns gene(gene symbol or geneid), zahard.ratio, zscore,
and p.value for each gene. The results list is used to construct weight vector
in the signature.
}
\usage{data("results")}
\format{
  The format is:
List of 3
 $ :'data.frame':	28 obs. of  4 variables:
  ..$ Gene        : chr [1:28] "CDKN1C" "CDKN2B" "DAB2IP" "DCC" ...
  ..$ Hazard.ratio: num [1:28] 1.027 0.685 0.663 0.494 0.634 ...
  ..$ zscore      : num [1:28] 0.0734 -1.162 -1.4084 -1.9798 -1.4282 ...
  ..$ p.value     : num [1:28] 0.9415 0.2452 0.159 0.0477 0.1532 ...
 $ :'data.frame':	26 obs. of  4 variables:
  ..$ Gene        : chr [1:26] "201335_s_at" "201621_at" "203185_at" "203525_s_at" ...
  ..$ Hazard.ratio: num [1:26] 2.712 0.849 0.742 0.99 0.402 ...
  ..$ zscore      : num [1:26] 1.7257 -0.3047 -0.6031 -0.0198 -1.6817 ...
  ..$ p.value     : num [1:26] 0.0844 0.7606 0.5464 0.9842 0.0926 ...
 $ :'data.frame':	34 obs. of  4 variables:
  ..$ Gene        : chr [1:34] "age" "gender" "sex" "cell_type" ...
  ..$ Hazard.ratio: num [1:34] NA NA NA NA NA ...
  ..$ zscore      : num [1:34] NA NA NA NA NA ...
  ..$ p.value     : num [1:34] NA NA NA NA NA ...
}

\examples{
data(results)
}
\keyword{datasets}
